Going to add platform

+++ Stage 2/RASTER.C
@@ -82,10 +82,10 @@ void plot_ball(UINT8 *base, int x, int y, UINT16 bitmap[8]) {
     }
 }
 
-void plot_brick (UINT8 *base, int x, int y, UINT16 bitmap[8]) {
+void plot_brick (UINT8 *base, int x, int y, UINT32 bitmap[8]) {
 	int row;
 	for (row = 0; row < 8; row++) {  /* iterate through rows */
-		UINT16 current_row = bitmap[row];
+		UINT32 current_row = bitmap[row];
         	UINT8 *pixel_ptr = base + (y + row) * (SCREEN_WIDTH / 8) + (x / 8);
 
 		int bit;
